+++ model/loginModel.php
@@ -1,0 +1,15 @@
+<?php
+require_once("connexion.php");
+
+function getUserByEmail($email) {
+    global $conn;
+
+    $sql = "SELECT * FROM users WHERE email = ?";
+    $stmt = mysqli_prepare($conn, $sql);
+    mysqli_stmt_bind_param($stmt, "s", $email);
+    mysqli_stmt_execute($stmt);
+    $result = mysqli_stmt_get_result($stmt);
+
+    return mysqli_fetch_assoc($result);
+}
+?>
